Added version 3.6.2165

+++ chromium/ui/platform_window/x11/atk_event_conversion.cc
@@ -7,6 +7,7 @@
 #include "base/check.h"
 #include "base/notreached.h"
 #include "ui/events/x/events_x_utils.h"
+#include "ui/gfx/x/connection.h"
 #include "ui/gfx/x/xproto.h"
 #include "ui/gfx/x/xproto_types.h"
 
@@ -25,12 +26,12 @@ std::unique_ptr<AtkKeyEventStruct> AtkKeyEventFromXEvent(
                             : ATK_KEY_EVENT_RELEASE;
 
   auto state = static_cast<int>(xkey->state);
-  auto keycode = static_cast<int>(xkey->detail);
+  auto keycode = xkey->detail;
   auto keysym = x11::Connection::Get()->KeycodeToKeysym(keycode, state);
 
   atk_key_event->state = state;
-  atk_key_event->keyval = static_cast<uint32_t>(keysym);
-  atk_key_event->keycode = keycode;
+  atk_key_event->keyval = keysym;
+  atk_key_event->keycode = static_cast<uint8_t>(keycode);
   atk_key_event->timestamp = static_cast<uint32_t>(xkey->time);
 
   // This string property matches the one that was removed from GdkEventKey. In
